requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
selenium
beautifulsoup4~=4.9.1
pandas~=1.3.1
pyautogui~=0.9.50
PyYAML~=5.3.1
lxml
future~=0.18.3
bs4~=0.0.1
future
python-dotenv
packaging